--- src/v1/index.ts
@@ -1,1 +1,0 @@
-export * from "./dijkstra";
